Refactor cells into multiple classes
